testheroku/requirements/common.txt
amqp==2.5.1
anyjson==0.3.3
billiard==3.5.0.5
boto3==1.9.210
botocore==1.12.210
celery==4.3.0
dj-static==0.0.6
Django==2.2.4
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-decouple==2.1
django-redis==4.10.0
django-storages==1.7.1
docutils==0.15.2
importlib-metadata==0.19
jmespath==0.9.4
kombu==4.2.1
Pillow==6.1.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
redis==2.10.6
s3transfer==0.2.1
six==1.12.0
sqlparse==0.3.0
static3==0.7.0
urllib3==1.25.3
vine==1.3.0
zipp==0.5.2
